# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PVT OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/contribution/list_months_import_contribution_command": {
            "post": {
                "tags": [
                    "IMPORTACION-APORTES-COMANDO"
                ],
                "summary": "LISTA LOS MESES QUE SE REALIZARON IMPORTACIONES A LA TABLA CONTRIBUTIONS DE COMANDO EN BASE A UN AÑO DADO EJ:2022",
                "description": "Lista los meses importados en la tabla contributions enviando como parámetro un año en específico",
                "operationId": "list_command_months",
                "requestBody": {
                    "description": "Provide auth credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "period_year": {
                                        "description": "Año de contribución a listar",
                                        "type": "integer",
                                        "example": "2022"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/contribution/import_contribution_command": {
            "post": {
                "tags": [
                    "IMPORTACION-APORTES-COMANDO"
                ],
                "summary": "PASO 3 IMPORTACIÓN DE CONTRIBUCION COMANDO GENERAL",
                "description": "Importación de aportes de Comando general",
                "operationId": "import_contribution_command",
                "requestBody": {
                    "description": "Provide auth credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "period_contribution": {
                                        "description": "fecha de aporte required",
                                        "type": "string",
                                        "example": "2022-03-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/contribution/report_import_contribution_command": {
            "post": {
                "tags": [
                    "IMPORTACION-APORTES-COMANDO"
                ],
                "summary": "GENERA REPORTE DE APORTES COMANDO IMPORTADAS",
                "description": "Genera reporte de aportes COMANDO de la tabla contribution de acuerdo a periodo de planilla",
                "operationId": "report_import_contribution_command",
                "requestBody": {
                    "description": "Provide auth credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "date_contribution": {
                                        "description": "fecha de planilla required",
                                        "type": "string",
                                        "example": "2022-03-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/contribution/list_years": {
            "get": {
                "tags": [
                    "METODOS-GLOBALES-IMPORTACION"
                ],
                "summary": "OBTIENE EL LISTADO DE AÑOS CONSECUTIVAMENTE ",
                "description": "Obtiene el listado de años de contribuciones de senasir de manera consecutiva hasta el año actual Ej 2022",
                "operationId": "list_years",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}